puller-js: isolates TestFile and TestName classes
